refactor(agents): extract orchestrator writeback to _orchestrator.writeback
Moves the live-DB writeback layer (apply_review_results_to_db),
the audit logger (_record_audit), the placeholder detector
(is_placeholder_description) + its _PLACEHOLDER_MARKERS sentinel,
the old-comment cache reader (_OldCommentReader), the per-row
SAVEPOINT context manager (_savepoint_or_passthrough), the
asset-label helper (_writeback_asset_label), and the live progress
factory (create_live_writeback_progress) out of orchestrator.py
into a focused amx/agents/_orchestrator/writeback.py.

Public surface re-exported from orchestrator.py so existing imports
(CLI run/history, web routers, analyze_flow, tests) keep working
unchanged. orchestrator.py: 1961 -> ~1450 LOC (-26%).
